requirements.txt
gradio
transformers
torch
pandas
scikit-learn
spacy
pymupdf
requests
biopython
bs4
pdfreader
tabula-py
spire.doc
Spire.XLS
thefuzz
wordsegment
spacy
spacy-lookups-data
gensim
xlrd>=2.0.1
openpyxl
gspread
oauth2client
nltk